requirements.txt
python-dotenv==1.0.1
python-telegram-bot==13.0
vk-api==11.9.9
redis==5.0.4